# SPARK-16785 R dapply doesn't return array or raw columns