--- a/R/pkg/R/DataFrame.R
+++ b/R/pkg/R/DataFrame.R
@@ -1381,8 +1381,9 @@ setMethod("dapplyCollect",
             # which is a serialized data.frame corresponds to one partition of the
             # SparkDataFrame.
             ldfs <- lapply(content, function(x) { unserialize(x[[1]]) })
-            ldf <- do.call(rbind, ldfs)
-            row.names(ldf) <- NULL
+# TODO Clark: a little experiment- comment out these two lines
+            #ldf <- do.call(rbind, ldfs)
+            #row.names(ldf) <- NULL
             ldf
           })