--- a/R/pkg/inst/worker/worker.R
+++ b/R/pkg/inst/worker/worker.R
@@ -27,21 +27,6 @@ elapsedSecs <- function() {
   proc.time()[3]
 }
 
-# Raws (binary data) require a list column in a data.frame
-# argument inputData should be a list of rows, with each row a list
-# Clark: Just for Testing
-#r = serialize(1, NULL)
-#inputData <- list(list(1, r), list(2, r), list(3, r))
-#rawcolumns <- c(FALSE, TRUE)
-
-rbind_with_raws <- function(inputData, rawcolumns){
-  listmatrix <- do.call(rbind, inputData)
-  # A dataframe with all list columns
-  out <- as.data.frame(listmatrix)
-  out[!rawcolumns] <- lapply(out[!rawcolumns], unlist)
-  out
-}
-
 compute <- function(mode, partition, serializer, deserializer, key,
              colNames, computeFunc, inputData) {
   if (mode > 0) {
@@ -52,17 +37,13 @@ compute <- function(mode, partition, serializer, deserializer, key,
       oldOpt <- getOption("stringsAsFactors")
       options(stringsAsFactors = FALSE)
 
-# TODO Clark: This didn't change error message
-# Handle binary data types
-row1 = inputData[[1]]
-rawcolumns <- ("raw" == sapply(row1, class))
-if(any(rawcolumns)){
-  inputData <- rbind_with_raws(inputData, rawcolumns)
-} else {
-  inputData <- do.call(rbind.data.frame, inputData)
-}
+      # Handle binary data types
+      if("raw" %in% sapply(inputData[[1]], class){
+        inputData <- SparkR:::rbind_with_raws(inputData)
+      } else {
+        inputData <- do.call(rbind.data.frame, inputData)
+      }
 
-      #inputData <- do.call(rbind.data.frame, inputData)
       options(stringsAsFactors = oldOpt)
 
       names(inputData) <- colNames