--- a/R/pkg/inst/worker/worker.R
+++ b/R/pkg/inst/worker/worker.R
@@ -39,7 +39,7 @@ compute <- function(mode, partition, serializer, deserializer, key,
 
       # Handle binary data types
       if("raw" %in% sapply(inputData[[1]], class){
-        inputData <- SparkR:::rbind_with_raws(inputData)
+        inputData <- SparkR:::rbindRaws(inputData)
       } else {
         inputData <- do.call(rbind.data.frame, inputData)
       }