@shivaram 2016-08-24
Thanks @clarkfitzg -- I'll take a look at this tomorrow