@clarkfitzg 2016-08-24
My pleasure. Let me know if / when I should squash these commits or rebase.

Working on some before and after benchmarks now.