@clarkfitzg 2016-08-25
Not completely sure though. I'll look into these timings a little further on Saturday to make sure I'm making a fair comparison.