@clarkfitzg 2016-08-29
Tried some more benchmarks today. Didn't see any difference in speed before / after patch. Observing the processes as they run I see the vast majority of time spent in the local R process, while just a couple seconds in the actual parallel evaluation of the functions.

![image](https://cloud.githubusercontent.com/assets/5356122/18046077/1b33331c-6e11-11e6-9128-de1e3e7870b2.png)