@felixcheung 2016-08-31
should we have a test against DataFrame with binary column?