@shivaram 2016-09-07
Sorry for the delay @clarkfitzg - The code change looks pretty good to me. I just had one question about mixed type columns.