@shivaram 2016-09-07
Thanks for the update. LGTM. Merging this to master and branch-2.0