@catlain 2017-06-13
done
[jira](https://issues.apache.org/jira/browse/SPARK-16785)